Add Test for Method NGram.get_suffixes
I'm realizing that the NGram.get_ngrams method does more work than the
user will want in many cases, so I'm going to make a method
NGram.get_suffixes that gives a mapping of suffixes to probabilities for
a given prefix. The internal procedures of NGram.get_ngrams will be
changed to utilize NGram.get_suffixes.