Add PoC Merge-Split Script
The script, which is a proof of concept, trains two markov models, each
based on a different source. The end goal is to generate texts that
sound distinctly like one source in parts, and distinctly like the other
source in other parts. This version achieves that goal to an extent.