Make merge_demo Scoop nlargest from Generator
merge_demo.py no longer relies on loading every generated iteration (of
potentially 1000s) into memory. Instead, it makes a generator and
collects the most erratic sequences into an n-sized heap.

NOTE: I have not at all convinced myself that this is actually working
as intended. In fact, from casual observation, I haven't seen any
speed-up. Granted, this might be because memory usage was never that
significant of a problem to begin with. Will have to check on this
later!