Make Markov End Random Sequence on Blank Token
The Markov model now ends its random sequence when it hits a suffix of
'' (i.e. a blank string). This is a problem because there is nothing in
the Markov model (as opposed to the TextMarkov model) that inherently
expects this. In other words this is a case of coupling / a lack of
separation of concerns. I'll have to do some work further down the line
to iron out the relationship between Markov and TextMarkov (they might
be flattened to one model).

Usage of the model had to be updated in one of the tests to reflect the
new definition of a 'sequence'.