Spin Off "Training" into its own Model
Instead of "modifying" a Markov "resource" using a special train_on
action, the client now is supposed to add a "training" object pointing
to the trainable model and the trainee. To be honest, I'm not sure how
much is gained by switching to this paradigm. It gels a little bit more
nicely with various forms of introspection, but it was really hard
writing the permissions for this style. Maybe I'll switch back later.